+++ hosts/hermes/darwin-configuration.nix
@@ -22,7 +22,7 @@
     # Update this when firefox-bin-unwrapped is merged
     (assert pkgs.firefox.meta.unsupported && pkgs.firefox-bin.meta.unsupported;
       "/Applications/Firefox.app")
-    "/System/Applications/Utilities/Terminal.app"
+    "/Applications/Ghostty.app"
     "/Applications/1Password.app"
     "${pkgs.vscode}/Applications/Visual Studio Code.app"
     "${pkgs.spotify}/Applications/Spotify.app"